requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud==0.34.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
pdf2image==1.16.3
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.8.2
